+++ Tests/TestUtils/Sourcery/GeneratedMocks/AutoMockable.generated.swift
@@ -2866,6 +2866,78 @@ class SynchronizerMock: Synchronizer {
         }
     }
 
+    // MARK: - lockMigrationResidual
+
+    var lockMigrationResidualAccountUUIDThrowableError: Error?
+    var lockMigrationResidualAccountUUIDCallsCount = 0
+    var lockMigrationResidualAccountUUIDCalled: Bool {
+        return lockMigrationResidualAccountUUIDCallsCount > 0
+    }
+    var lockMigrationResidualAccountUUIDReceivedAccountUUID: AccountUUID?
+    var lockMigrationResidualAccountUUIDReturnValue: Zatoshi!
+    var lockMigrationResidualAccountUUIDClosure: ((AccountUUID) async throws -> Zatoshi)?
+
+    func lockMigrationResidual(accountUUID: AccountUUID) async throws -> Zatoshi {
+        if let error = lockMigrationResidualAccountUUIDThrowableError {
+            throw error
+        }
+        lockMigrationResidualAccountUUIDCallsCount += 1
+        lockMigrationResidualAccountUUIDReceivedAccountUUID = accountUUID
+        if let closure = lockMigrationResidualAccountUUIDClosure {
+            return try await closure(accountUUID)
+        } else {
+            return lockMigrationResidualAccountUUIDReturnValue
+        }
+    }
+
+    // MARK: - unlockMigrationResidual
+
+    var unlockMigrationResidualAccountUUIDThrowableError: Error?
+    var unlockMigrationResidualAccountUUIDCallsCount = 0
+    var unlockMigrationResidualAccountUUIDCalled: Bool {
+        return unlockMigrationResidualAccountUUIDCallsCount > 0
+    }
+    var unlockMigrationResidualAccountUUIDReceivedAccountUUID: AccountUUID?
+    var unlockMigrationResidualAccountUUIDReturnValue: Int!
+    var unlockMigrationResidualAccountUUIDClosure: ((AccountUUID) async throws -> Int)?
+
+    func unlockMigrationResidual(accountUUID: AccountUUID) async throws -> Int {
+        if let error = unlockMigrationResidualAccountUUIDThrowableError {
+            throw error
+        }
+        unlockMigrationResidualAccountUUIDCallsCount += 1
+        unlockMigrationResidualAccountUUIDReceivedAccountUUID = accountUUID
+        if let closure = unlockMigrationResidualAccountUUIDClosure {
+            return try await closure(accountUUID)
+        } else {
+            return unlockMigrationResidualAccountUUIDReturnValue
+        }
+    }
+
+    // MARK: - estimateMigrationRuns
+
+    var estimateMigrationRunsAccountUUIDThrowableError: Error?
+    var estimateMigrationRunsAccountUUIDCallsCount = 0
+    var estimateMigrationRunsAccountUUIDCalled: Bool {
+        return estimateMigrationRunsAccountUUIDCallsCount > 0
+    }
+    var estimateMigrationRunsAccountUUIDReceivedAccountUUID: AccountUUID?
+    var estimateMigrationRunsAccountUUIDReturnValue: MigrationRunEstimate!
+    var estimateMigrationRunsAccountUUIDClosure: ((AccountUUID) async throws -> MigrationRunEstimate)?
+
+    func estimateMigrationRuns(accountUUID: AccountUUID) async throws -> MigrationRunEstimate {
+        if let error = estimateMigrationRunsAccountUUIDThrowableError {
+            throw error
+        }
+        estimateMigrationRunsAccountUUIDCallsCount += 1
+        estimateMigrationRunsAccountUUIDReceivedAccountUUID = accountUUID
+        if let closure = estimateMigrationRunsAccountUUIDClosure {
+            return try await closure(accountUUID)
+        } else {
+            return estimateMigrationRunsAccountUUIDReturnValue
+        }
+    }
+
     // MARK: - signAndStoreMigrationSchedule
 
     var signAndStoreMigrationScheduleAccountUUIDUskThrowableError: Error?